move secret directory
