+++ src/common/modules/caches/index.ts
@@ -2,7 +2,7 @@ import { CacheModule, Module } from '@nestjs/common';
 import { CachesService } from './caches.service';
 import * as redisStore from 'cache-manager-redis-store';
 import { GCloudSecretManagerService } from '@debionetwork/nestjs-gcloud-secret-manager';
-import { keyList } from '../../../secrets';
+import { keyList } from '../../secrets';
 
 @Module({
   imports: [
