+++ src/common/modules/health-indicators/elasticsearch.health/index.ts
@@ -3,7 +3,7 @@ import { Module } from '@nestjs/common';
 import { ElasticsearchModule } from '@nestjs/elasticsearch';
 import { TerminusModule } from '@nestjs/terminus';
 import { ElasticsearchHealthIndicator } from './elasticsearch.health.indicator';
-import { keyList } from '../../../../secrets';
+import { keyList } from '../../../secrets';
 
 @Module({
   imports: [
