+++ src/endpoints/transaction/transaction.module.ts
@@ -1,7 +1,7 @@
 import { GCloudSecretManagerService } from '@debionetwork/nestjs-gcloud-secret-manager';
 import { Module } from '@nestjs/common';
 import { ElasticsearchModule } from '@nestjs/elasticsearch';
-import { keyList } from '../../secrets';
+import { keyList } from '../../common/secrets';
 import { TransactionLoggingModule } from '../../common';
 import { TransactionController } from './transaction.controller';
 import { TransactionService } from './transaction.service';
